+++ packages/strapi-plugin-content-type-builder/admin/src/containers/FormModal/utils/attributes/advancedForm.js
@@ -33,12 +33,12 @@ const advancedForm = {
       ],
     };
   },
-  component: ({ isRepeatable }, step) => {
+  component: ({ repeatable }, step) => {
     if (step === '1') {
       return { items: componentForm.advanced('componentToCreate') };
     }
 
-    if (isRepeatable) {
+    if (repeatable) {
       return { items: [[options.required], [uiHelpers.divider], [options.max], [options.min]] };
     }
