+++ packages/strapi-plugin-content-type-builder/admin/src/containers/FormModal/utils/attributes/commonBaseForm.js
@@ -1,0 +1,7 @@
+import { nameField } from './constants';
+
+const commonBaseForm = {
+  items: [[nameField]],
+};
+
+export default commonBaseForm;
