+++ packages/strapi-plugin-content-type-builder/admin/src/containers/FormModal/utils/attributes/constants.js
@@ -1,0 +1,19 @@
+import getTrad from '../../../../utils/getTrad';
+
+const nameField = {
+  autoFocus: true,
+  name: 'name',
+  type: 'text',
+  label: {
+    id: getTrad('modalForm.attribute.form.base.name'),
+  },
+  description: {
+    id: getTrad('modalForm.attribute.form.base.name.description'),
+  },
+  validations: {
+    required: true,
+  },
+};
+
+// eslint-disable-next-line import/prefer-default-export
+export { nameField };
