+++ packages/strapi-plugin-content-type-builder/admin/src/containers/FormModal/utils/attributes/form.js
@@ -1,7 +1,9 @@
 import advancedForm from './advancedForm';
+import baseForm from './baseForm';
 
 const form = {
   advanced: advancedForm,
+  base: baseForm,
 };
 
 export default form;
