+++ packages/strapi-plugin-content-type-builder/admin/src/containers/FormModal/utils/attributes/index.js
@@ -1,2 +1,3 @@
 // eslint-disable-next-line import/prefer-default-export
 export { default as attributesForm } from './form';
+export { default as commonBaseForm } from './commonBaseForm';
