+++ packages/strapi-plugin-content-type-builder/admin/src/containers/FormModal/utils/attributes/uiHelpers.js
@@ -2,6 +2,10 @@ const uiHelpers = {
   divider: {
     type: 'divider',
   },
+  spacer: {
+    type: 'spacer',
+  },
+  spacerMedium: { type: 'spacer-medium' },
 };
 
 export default uiHelpers;
