+++ packages/strapi-plugin-content-type-builder/admin/src/containers/FormModal/utils/component/index.js
@@ -1,3 +1,4 @@
 // eslint-disable-next-line import/prefer-default-export
 export { default as createComponentSchema } from './createComponentSchema';
+export { default as componentField } from './componentField';
 export { default as componentForm } from './form';
