+++ packages/strapi-plugin-content-type-builder/admin/src/containers/FormModal/utils/dynamicZone/form.js
@@ -1,28 +1,28 @@
 import getTrad from '../../../../utils/getTrad';
 
-import { componentForm } from '../component';
+import { componentField, componentForm } from '../component';
 
-const componentField = {
-  label: {
-    id: getTrad('modalForm.attribute.text.type-selection'),
-  },
-  name: 'createComponent',
-  type: 'booleanBox',
-  size: 12,
-  options: [
-    {
-      headerId: getTrad('form.attribute.component.option.create'),
-      descriptionId: getTrad('form.attribute.component.option.create.description'),
-      value: true,
-    },
-    {
-      headerId: getTrad('form.attribute.component.option.reuse-existing'),
-      descriptionId: getTrad('form.attribute.component.option.reuse-existing.description'),
-      value: false,
-    },
-  ],
-  validations: {},
-};
+// const componentField = {
+//   label: {
+//     id: getTrad('modalForm.attribute.text.type-selection'),
+//   },
+//   name: 'createComponent',
+//   type: 'booleanBox',
+//   size: 12,
+//   options: [
+//     {
+//       headerId: getTrad('form.attribute.component.option.create'),
+//       descriptionId: getTrad('form.attribute.component.option.create.description'),
+//       value: true,
+//     },
+//     {
+//       headerId: getTrad('form.attribute.component.option.reuse-existing'),
+//       descriptionId: getTrad('form.attribute.component.option.reuse-existing.description'),
+//       value: false,
+//     },
+//   ],
+//   validations: {},
+// };
 
 const form = {
   advanced: {
